add lazy loading and loading screen
